backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
PyMuPDF==1.25.5
pyxirr==0.10.8
mistralai>=1.2.0
pydantic>=2.9.0
python-dotenv>=1.0.1
